moved task actions into task epic
