+++ frontend/src/constants/ProjectActionTypes.js
@@ -10,19 +10,15 @@ export const CHECK_STAKED_STATUS = 'CHECK_STAKED_STATUS'
 // export const PROJECT_STATE_RECEIVED = 'PROJECT_STATE_RECEIVED'
 export const CHECK_ACTIVE_STATUS = 'CHECK_ACTIVE_STATUS'
 export const ACTIVE_STATUS_CHECKED = 'ACTIVE_STATUS_CHECKED'
+
 export const SET_TASK_LIST = 'SET_TASK_LIST'
 export const SUBMIT_HASHED_TASK_LIST = 'SUBMIT_HASHED_TASK_LIST'
 export const TASK_LIST_SET = 'TASK_LIST_SET'
-export const TASK_COMPLETED = 'TASK_COMPLETED'
-export const TASK_VALIDATED = 'TASK_VALIDATED'
 export const STAKED_STATUS_CHECKED = 'STAKED_STATUS_CHECKED'
 export const HASHED_TASK_LIST_SUBMITTED = 'HASHED_TASK_LIST_SUBMITTED'
-export const FINAL_TASK_LIST_SUBMITTED = 'FINAL_TASK_LIST_SUBMITTED'
-export const SUBMIT_FINAL_TASK_LIST = 'SUBMIT_FINAL_TASK_LIST'
 export const GET_VERIFIED_TASK_LISTS = 'GET_VERIFIED_TASK_LISTS'
 export const VERIFIED_TASK_LISTS_RECEIVED = 'VERIFIED_TASK_LISTS_RECEIVED'
-export const CLAIM_TASK = 'CLAIM_TASK'
-export const TASK_CLAIMED = 'TASK_CLAIMED'
+
 // export const UPDATE_PROJECT = 'UPDATE_PROJECT'
 
 export const GET_PROJECTS = 'GET_PROJECTS'
