+++ frontend/src/constants/TaskActionTypes.js
@@ -1,0 +1,6 @@
+export const FINAL_TASK_LIST_SUBMITTED = 'FINAL_TASK_LIST_SUBMITTED'
+export const SUBMIT_FINAL_TASK_LIST = 'SUBMIT_FINAL_TASK_LIST'
+export const CLAIM_TASK = 'CLAIM_TASK'
+export const TASK_CLAIMED = 'TASK_CLAIMED'
+export const TASK_COMPLETED = 'TASK_COMPLETED'
+export const TASK_VALIDATED = 'TASK_VALIDATED'
